webauth/requirements.txt
requests
python-jose==3.2.0